+++ src/scripts/background/chrome/icon/clickCount.js
@@ -5,6 +5,14 @@ import {getStorageData} from "../../../shared/utils/chromeStorage";
 // Array of click counts for each modal
 export const clickCounts = [14, 25, 40, 40] //[20, 40, 60, 80, 100];
 
+export function initClickCount() {
+  chrome.runtime.onInstalled.addListener(
+    (details) => {
+      initClickIconCount(details);
+    }
+  );
+}
+
 export function initClickIconCount(details) {
   if (details.reason === "install" || details.reason === "update") {
     chrome.storage.sync.set({"clickIconCount": clickCounts[0], "modalIndex": 0});
@@ -20,4 +28,3 @@ export async function updateClickIconCount() {
 export function resetClickCount(count) {
   chrome.storage.sync.set({"clickIconCount": count}); //100
 }
-
