+++ src/scripts/background/chrome/icon/iconAction.js
@@ -1,4 +1,13 @@
-export async function callContentScript(tab) {
+/***
+ * This is the code for the extension to run when the icon is clicked
+ */
+export function listenIconClick() {
+  chrome.action.onClicked.addListener(async (tab) => {
+    await launchIconClickAction(tab);
+  });
+}
+
+export async function launchIconClickAction(tab) {
   if (!tab || tab.url.startsWith("chrome://")) {
     console.info(`Tab ${tab?.id || ''} is not injectable`);
     return;
