--- src/scripts/background/modals/defineDisplayModalOnInstall.js
@@ -1,14 +1,0 @@
-export function defineDisplayModalOnInstall(details) {
-  switch (details.reason) {
-    case "install":
-      chrome.storage.sync.set({displayModalWelcome: true},
-        () => console.log("Welcome modal will be displayed")
-      );
-      break;
-    case "update":
-      chrome.storage.sync.set({displayModalUpdate: true},
-        () => console.log("Last update modal will be displayed")
-      );
-      break;
-  }
-}
